requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyJWT[crypto]==2.8.0
httpx==0.27.0